requirements.txt
obsws-python
Pillow
opencv-python
numpy
keyboard
python-dotenv
